+++ src/main/java/fr/adaming/forum/dao/IAddressDao.java
@@ -5,7 +5,7 @@ import fr.adaming.forum.entity.Address;
 public interface IAddressDao {
 		
 		public Address addAdresse(Address address);
-		public Address getAdresseById(int idAddress);
+		public Address getAdresseById(Long idAddress);
 		public Address updateAdresse(Address address);
-		public Address deleteAdresse(int idAddress);
+		public Address deleteAdresse(Long idAddress);
 }
