+++ src/main/java/fr/adaming/forum/service/IAddressService.java
@@ -5,8 +5,8 @@ import fr.adaming.forum.entity.Address;
 public interface IAddressService {
 	
 	public Address addAdresse(Address address);
-	public Address getAdresseById(int idAddress);
+	public Address getAdresseById(Long idAddress);
 	public Address updateAdresse(Address address);
-	public Address deleteAdresse(int idAddress);
+	public Address deleteAdresse(Long idAddress);
 
 }
