+++ src/main/java/fr/adaming/forum/service/IFormationService.java
@@ -7,9 +7,9 @@ import fr.adaming.forum.entity.Formation;
 public interface IFormationService{
 	
 	public Formation addFormation(Formation formation);
-	public Formation getFormationById(int idFormation);
+	public Formation getFormationById(Long idFormation);
 	public Formation updateFormation(Formation formation);
-	public Formation deleteFormation(int idFormation);
+	public Formation deleteFormation(Long idFormation);
 	public List<Formation>  getAllFormations();
 	public List<Formation>  getFormationByKeyWord(String keyWord);
