+++ src/main/java/fr/adaming/forum/service/ISkillService.java
@@ -8,6 +8,6 @@ public interface ISkillService {
 	public Skill addSkill(Skill skill);
 	public List<Skill> getSkillByKeyWord(String keyWord);
 	public Skill updateSkill(Skill skill);
-	public Skill deleteSkill(int idSkill);
+	public Skill deleteSkill(Long idSkill);
 	public List<Skill> getAllSkills ();
 }
